# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Rural Services API
  description: 'This is a Rural Services module. You can find out more about Rural Services at [http://ruralservices.io](http://ruralservices.io) or on [www.ruralservices.gov.in, #ruralservices](http://swagger.io/irc/).'
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Rural Services
  url: http://swagger.io
servers:
  - url: https://localhost:8080/v1/api/nxg-ecomm
  - url: http://localhost:8080/v1/api/nxg-ecomm
tags:
  - name: Rural Services API
    description: Everything about Rural Services
    externalDocs:
      description: Find out more
      url: http://swagger.io
paths:
  /applicantinformation:
    post:
      tags:
        - Applicant Information
      summary: Adding a new applicant
      requestBody:
        description: Applicant details that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicantInformation'
        required: true
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body
  /applicantinformation/familymember/:
    post:
      tags:
        - Applicant Information Member
      summary: Adding a new member for the applicant
      requestBody:
        description: Member details that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicantFamily'
        required: true
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body
    get:
      tags:
        - Applicant Information Member
      summary: Adding a new member for the applicant
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body

  /applicantinformation/familymember/{memberId}:
    put:
      tags:
        - Applicant Information Member
      summary: Adding a new member for the applicant
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: Member details that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicantFamily'
        required: true
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body
    delete:
      tags:
        - Applicant Information Member
      summary: Adding a new member for the applicant
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Successful
          content: {}      
        '400':
          description: Bad request input
          content: {}
        '404':
          description: Data not found
          content: {}
        '500':
          description: Invalid input
          content: {}          
        '502':
          description: Bad Gatway
          content: {}               
      x-codegen-request-body-name: body
      
components:
  schemas:
    ApplicantInformation:
      type: object
      required:
       - karyaSwaroop
       - karyaPrakriti1
       - name
       - nameasAADHAR
       - fatherNameorHusbandName
       - fatherORhusband
       - motherName
       - gender
       - marriage
       - dob
       - age
       - caste
       - ration
       - voterId
       - rsBy
       - aadharCard
       - bankName
       - branchName
       - acNumber
       - ifsCode
       - esiNumber
       - epfNumber
      properties:
        pmjjyMoney:
          type: boolean
        karyaSwaroop:
          type: integer
          maxLength: 2
        karyaPrakriti1:
          type: integer
          maxLength: 3
        karyaPrakriti2:
          type: integer
          maxLength: 3
        karyaPrakriti3:
          type: integer       
          maxLength: 3
        programName:
          type: integer
          maxLength: 3
        name:
          type: string
          maxLength: 120
        nameasAADHAR:
          type: string
          maxLength: 120
        familyHeadMemberName:
          type: string
          maxLength: 120
        fatherNameorHusbandName:
          type: string
          maxLength: 120
        fatherORhusband:
          type: integer       
          maxLength: 1
        motherName:
          type: string
          maxLength: 120
        gender:
          type: integer       
          maxLength: 1
        marriage:
          type: integer       
          maxLength: 1
        dob:
          type: string
          maxLength: 10
        age:
          type: integer       
          maxLength: 3
        caste:
          type: integer       
          maxLength: 1
        manrega:
          type: boolean
        ration:
          type: string
          maxLength: 50
        voterId:
          type: string
          maxLength: 50
        rsBy:
          type: string
          maxLength: 50
        aadharCard:
          type: string
          maxLength: 50
        bankName:
          type: string
          maxLength: 50
        branchName:
          type: string
          maxLength: 50
        acNumber:
          type: string
          maxLength: 25
        ifsCode:
          type: string
          maxLength: 15
        esiNumber:
          type: string
          format: varchar
          maxLength: 12
        epfNumber:
          type: string
          format: varchar
          maxLength: 12
        sameAsCurrentAddress:
          type: boolean
        currentAddress:
          $ref: '#/components/schemas/ApplicantCurrentAddress'
        permanentAddress:
          $ref: '#/components/schemas/ApplicantPermanentAddress'
        applicantId:
          type: integer
          format: bigint
          maxLength: 10
    ApplicantAddress:
      type: object
      required:
       - district
       - gramin
       - shehri
       - vidhansabhaArea
       - vikasKhand
       - panchayat
       - ward
       - address
       - houseNo
      properties:
        district:
          type: integer       
          maxLength: 3
        selectedAddressType:    
          type: integer
        vidhansabhaArea:
          type: integer       
          maxLength: 3
        vikasKhand:
          type: integer       
          maxLength: 3
        panchayat:
          type: integer
          maxLength: 3
        ward:
          type: integer
          maxLength: 3
        address:
          type: string
          maxLength: 200
        houseNo:
          type: string
          maxLength: 50
        mobile1:
          type: string
          maxLength: 10
        mobile2:
          type: string
          maxLength: 10
    ApplicantCurrentAddress:
      $ref: '#/components/schemas/ApplicantAddress'
    ApplicantPermanentAddress:
      $ref: '#/components/schemas/ApplicantAddress'
    ApplicantFamily:
      type: object
      properties: 
        applicantId:
          type: integer
          format: bigint
          maxLength: 10
        members:
          type: array
          required:
          - memberId
          - memberName
          - memberDOB
          - age
          - memberGender
          items:
            type: object
            properties:
              memberId: 
                type: integer
                format: bigint
                maxLength: 10
              memberName:
                type: string
              relation:
                type: integer  
              memberDOB:
                type: string
              age:
                type: integer
              memberGender:
                type: integer
              memberAadhar:
                type: string
              nominee:
                type: boolean
              memberPercentDistribution:
                type: integer 
              memberEducation:
                type: integer 
              memberSchoolGrade:
                type: integer 
              memberMaritalStatus:
                type: integer 
              programName:  
                type: integer 
              status:  
                type: integer 